# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Absensi",
        "description": "API Aplikasi absensi berbasis web (e-absensi) online untuk memantau kinerja karyawan jarak jauh, dan pekerja work from home (WFH).",
        "termsOfService": "http://127.0.0.1:8000/terms/",
        "contact": {
            "email": "bandungcoders@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "OpenApi host"
        }
    ],
    "paths": {
...
        "/password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change password",
                "description": "",
                "operationId": "password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/form-data": {
                            "schema": {
                                "properties": {
                                    "old_password": {
                                        "description": "Old password user",
                                        "type": "string"
                                    },
                                    "new_password": {
                                        "description": "New password",
                                        "type": "string"
                                    },
                                    "new_confirm_password": {
                                        "description": "password confirm",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}